moving vamc police to products
