+++ .github/CODEOWNERS
@@ -20,3 +20,4 @@ src/products/storyListing/ @department-of-veterans-affairs/ap-team
 src/products/pressRelease/ @department-of-veterans-affairs/ap-team
 src/products/pressReleaseListing/ @department-of-veterans-affairs/ap-team
 src/products/leadershipListing/ @department-of-veterans-affairs/ap-team
+src/products/vamcSystemVaPolice/ @department-of-veterans-affairs/ap-team
