+++ src/data/queries/index.ts
@@ -46,7 +46,7 @@ import * as SupportServices from '@/data/queries/supportServices'
 import * as Table from '@/data/queries/table'
 import * as VamcEhr from '@/data/queries/vamcEhr'
 import * as VamcSystem from '@/data/queries/vamcSystem'
-import * as VamcSystemVaPolice from '@/data/queries/vamcSystemVaPolice'
+import * as VamcSystemVaPolice from '@/products/vamcSystemVaPolice/query'
 import * as VetCenter from '@/data/queries/vetCenter'
 import * as VetCenterHealthServices from '@/data/queries/vetCenterHealthServices'
 import * as VetCenterOutstation from '@/data/queries/vetCenterOutstation'
