+++ src/products/vamcSystemVaPolice/formatted-type.ts
@@ -1,9 +1,9 @@
-import { PublishedEntity } from './publishedEntity'
+import { PublishedEntity } from '@/types/formatted/publishedEntity'
 import { SideNavMenu } from '@/types/formatted/sideNav'
 import { Wysiwyg } from '@/types/formatted/wysiwyg'
 import { PhoneNumber } from '@/types/formatted/phoneNumber'
-import { FeaturedContent } from './featuredContent'
-import { QaSection } from './qaSection'
+import { FeaturedContent } from '@/types/formatted/featuredContent'
+import { QaSection } from '@/types/formatted/qaSection'
 import { LovellChildVariant } from '@/lib/drupal/lovell/types'
 
 export type VamcSystemVaPolice = PublishedEntity & {
